Aula3

Description

Este módulo contém definições Haskell para o cálculo de funções recursivas simples (obs: isto é apenas uma descrição mais longa do módulo para efeitos de documentação...).

Documentation

adicionaLista

Arguments

:: [Int]
Lista de Inteiros.

-> Int
Valor a ser adicionado a cada elemento.

-> [Int]
Lista resultante após adicionar o valor.

A função adicionaLista recebe uma lista de inteiros e um número inteiro n, e retorna uma nova lista onde cada elemento da lista original foi somado a n.

Exemplos de utilização:

>>> adicionaLista [1, 2, 3, 4] 5
[6,7,8,9]
>>> adicionaLista [] 10
[]

removeString

Arguments

:: [String]
Lista de Strings.

-> Char
Caracter a verificar e a remover no caso da String começar por esse caractere.

-> [String]
Lista resultante da remoção das strings iniciadas pelo caratere anterior.

A função removeString recebe uma lista de strings e um caractere. Ela retorna uma nova lista de strings onde todas as strings que começam com o caractere dado foram removidas.

Exemplos de utilização:

>>> removeString ["abc", "def", "ghi", "jkl"] 'd'
["abc","ghi","jkl"]
>>> removeString ["abacaxi", "banana", "pessego", "maracuja"] 'a'
["banana","pessego","maracuja"]

somaPar

Arguments

:: [(Int, Int)]
Lista de pares de inteiros.

-> Int
Valor inteiro a adicionar à primeira componente de cada par de inteiros.

-> [(Int, Int)]
Lista de pares de inteiros resultante da adição do Inteiro à primeira componente de cada par.

A função somaPar recebe uma lista de pares de inteiros e um valor n. Ela retorna uma nova lista de pares onde n foi adicionado à primeira componente de cada par.

Exemplos de utilização:

>>> somaPar [(1, 2), (3, 4), (5, 6)] 10
[(11,2),(13,4),(15,6)]
>>> somaPar [] 5
[]

maiorSegundaComponente

Arguments

:: [(Int, Int)]
Lista de Pares de Inteiros

-> Int
Valor inteiro que corresponde ao maior valor encontrado na segunda componente dos pares.

A função maiorSegundaComponente recebe uma lista não vazia de pares de inteiros. Ela retorna o maior valor encontrado na segunda componente dos pares.

Exemplos de utilização:

>>> maiorSegundaComponente [(1, 5), (3, 9), (2, 7)]
9

nextDigit

Arguments

:: Char
Dígito.

-> Char
Dígito seguinte.

A função ´proxdig´ é uma função que recebe um digíto e calcula o próximo dígito

Exemplo de utilização:

>>> proxdig 2
3

Propriedades:

proxdig a

| ord a == 57 = '0'

| otherwise = chr (ord a + 1)

proxdigl

Arguments

:: [Char]
Lista de digítos

-> [Char]
Lista de dígitos transformada

A função ´proxdigl´ é uma função que recebe uma lista de digítos e substitui cada um deles pelo seu sucessor, usando a função anterior

Exemplo de utilização:

>>> proxdigl ['1','2','3']
"234"

Propriedades:

proxdigl [] = []

proxdigl (a:as) = (proxdig (head a)) : proxdigl as

nextVogal :: Char -> Char

A função nextVogal recebe um caractere e retorna a próxima vogal após esse caractere, considerando que a próxima vogal após u é a. Caso o caractere não seja uma vogal (aeiou), a função mantém o caractere original.

Exemplo de Uso

>>> nextVogal 'e'
'i'
>>> nextVogal 'i'
'o'

Propriedades

nextVogal 'a' = 'e'

nextVogal 'e' = 'i'

nvogal

Arguments

:: String
Lista de vogais

-> String
Lista de vogais tranformada

A função ´nvogal´ é uma função que recebe uma lista de vogais e substitui cada uma delas pela vogal seguinte

Exemplo de utilização:

>>> proxdigl ['a','e','i']
"eio"

Propriedades:

nvogal [] = []

nvogal (c:cs)

|c == 'a' = 'e' : nvogal cs

| c == 'e' = 'i' : nvogal cs

| c == 'i' = 'o' : nvogal cs

| c == 'o' = 'u' : nvogal cs

| c == 'u' = 'a' : nvogal cs

type Nome = String

type Coordenada = (Int, Int)

data Movimento

Constructors

N

S

E

W

type Movimentos = [Movimento]

data PosicaoPessoa

Constructors

Pos Nome Coordenada

posicao

Arguments

:: PosicaoPessoa
Posição inicial.

-> Movimentos
Sentido dos movimentos

-> PosicaoPessoa
Posição após ocorrer movimento

A função posicao calcula a posição de uma pessoa depois de executar uma sequência de movimentos:

posicao :: PosicaoPessoa
             -> Movimentos
             -> PosicaoPessoa
posicao (Pos n (x,y)) [] = Pos n (x,y)
posicao (Pos n (x,y)) (m:ms) =
    case m of
        N -> posicao (Pos n (x,y+1)) ms
        S -> posicao (Pos n (x, y-1)) ms
        E -> posicao (Pos n (x+1,y)) ms 
        W -> posicao (Pos n (x-1,y)) ms

Exemplos de utilização:

>>> posicao (Pos "Josuke" (1,2)) [N,E]

posicoesM

Arguments

:: [PosicaoPessoa]
Lista de Posições de pessoas

-> Movimento
Movimento a realizar

-> [PosicaoPessoa]
Lista de Posições de pessoas após movimento

A função posicoesM é uma função que dada uma lista de posições de pessoas, actualiza essa lista depois de todas executar um movimento dado:

Exemplo de utilização:

>>> posicoesM [Pos "Joel" (0,0),Pos "Ellie" (3,2),Pos "Abby" (2,6)] N
[Pos "Joel" (0,1),Pos "Ellie" (3,3),Pos "Abby" (2,7)]

Propriedades:

posicoesM [] _ = []

posicoesM (Pos n (x, y):rn) m =

case m of

N -> Pos n (x, y + 1) : posicoesM rn m -- Norte (aumenta y)

S -> Pos n (x, y - 1) : posicoesM rn m -- Sul (diminui y)

E -> Pos n (x + 1, y) : posicoesM rn m -- Leste (aumenta x)

W -> Pos n (x - 1, y) : posicoesM rn m -- Oeste (diminui x)

posicoesMs

Arguments

:: [PosicaoPessoa]
Lista de Posições de pessoas inicial

-> Movimentos
Sequência de movimentos a ser executada por todas as pessoas

-> [PosicaoPessoa]
Lista de Posições de pessoas após os movimentos

A função posicoesMs é uma função que dada uma lista de posições de pessoas, actualiza essa lista depois de todas as pessoas executarem uma mesma sequência de movimentos. Esta função usa as funções anteriormente definidas.

Exemplo de utilização:

>>> posicoesMs [Pos "Joel" (0,0),Pos "Ellie" (3,2),Pos "Abby" (2,6)] [N,E]
[Pos "Joel" (1,1),Pos "Ellie" (4,3),Pos "Abby" (3,7)]

Propriedades:

posicoesMs [] _ = []

posicoesMs (p:ps) m =

let pa = posicao p m

in  pa : posicoesMs ps m

pessoasNorte

Arguments

:: [PosicaoPessoa]
Lista de pessoas e respetivas posições

-> [Nome]
Nome(s) de pessoa(s) mais a Norte

A função pessoasNorte é uma função que calcula quais são as pessoas posicionadas mais a norte.

Exemplo de utilização:

>>> pessoasNorte [Pos "Joel" (0,0),Pos "Ellie" (3,2),Pos "Abby" (2,6)]
"Abby"

Propriedades:

pessoasNorte [] = []

pessoasNorte posicoes = aux posicoes [] maxCoordY

where

aux :: [PosicaoPessoa] -> [Nome] -> Int -> [Nome]

aux [] nomes _ = nomes

aux ((Pos nome (x, y)):resto) nomes maxY

| y > maxY = aux resto [nome] y

| y == maxY = aux resto (nome : nomes) maxY

| otherwise = aux resto nomes maxY

maxCoordY :: Int

maxCoordY = maximum [y | (Pos _ (_, y)) <- posicoes]

deslocaEsquerda :: [a] -> [a]

A função dlista é uma função que recebe uma lista e desloca cada elemento da lista, n posições para a direita.

Exemplo de utilização:

>>> dlista 2 [1,2,3,4,5]
[4,5,1,2,3]

Propriedades:

dlista _ [] = []

dlista n lista =

let tamanho = length lista

deslocamento = n `mod` tamanho

(inicio, fim) = splitAt (tamanho - deslocamento) lista

in fim ++ inicio

deslocaEsquerdaN :: [a] -> Int -> [a]

deslocaDireita :: [a] -> [a]

Desloca uma posição para a direita

deslocaDireitaN :: [a] -> Int -> [a]

>>> deslocaDireitaN [1,2,3,4] 2
[3,4,1,2]

dlista

Arguments

:: Int
Numero de posições a serem deslocadas

-> [a]
Lista inicial

-> [a]
Lista após deslocamento

elista

Arguments

:: Int
Número de posições a serem deslocadas para a esquerda

-> [a]
Lista inicial

-> [a]
Lista após o deslocamento

A função elista é uma função que recebe uma lista e desloca cada elemento da lista, n posições para a esquerda.

Exemplo de utilização:

>>> dlista 2 [1,2,3,4,5]
[3,4,5,1,2]

Propriedades:

dlista _ [] = []

dlista n lista =

let tamanho = length lista

deslocamento = n `mod` tamanho

(inicio, fim) = splitAt deslocamento lista

in fim ++ inicio