Main

Description

Módulo para a realização da Tarefa 6 de LI1 em 2023/24.

Documentation

data Menu

Neste módulo realizamos as funções necessárias para pôr a parte gáfica a funcionar. Para isso, tivemos de definir alguns data Types novos:

Menu.

Constructors

InGame

MenuInicial

MenuCreditos

MenuPausa

Win

Lose

Options

data Opcoes

Opcoes.

Constructors

Default

Jogar

Sair

Creditos

Voltar

Continuar

Opcoes

type Pictures = [Picture]

Pictures.

data TipoImagens

Tipo de imagens.

Constructors

ImageBloc

ImagePer

ImageIni

type ImagensB = [(Bloco, Picture)]

type ImagensP = [(Temas, [(Entidade, Pictures)])]

type ImagensC = [(Colecionavel, Picture)]

type ImagensM = [(Menu, Pictures)]

data Imagens

Constructors

Imagens

Fields

blocos :: ImagensB

personagens :: ImagensP

colecionaveis :: ImagensC

menus :: ImagensM

data Temas

Temas

Constructors

Escuro

Claro

data PrimateKong

Data type do Jogo.

Constructors

PK

Fields

jogo :: Jogo

menu :: Menu

opcoes :: Opcoes

imagens :: Imagens

tema :: Temas

tempo :: Tempo

semente :: Semente

paridadet :: ParidadeT

carteclas :: CarregaTecla

mainDisplay :: Display

Definições da janela do jogo.

corfundo :: Color

Cor do fundo.

fr :: Int

O framerate do jogo.

geraSeed :: IO Int

Seed do jogo.

estadoInicial :: IO PrimateKong

Estado inicial do jogo.

gImageBloco :: IO ImagensB

Agrupa as imagens dos blocos nas suas diferentes categorias.

gImagePersonagem :: IO ImagensP

Agrupa as imagens dos Personagens nas suas diferentes categorias.

gImageColec :: IO ImagensC

Agrupa as imagens dos colecionaveis nas suas diferentes categorias.

gImageMenu :: IO ImagensM

Agrupa as imagens dos menus nas suas diferentes categorias.

desenhaEstado :: PrimateKong -> Picture

Desenha Estado do Jogo.

desenhaMenu :: Menu -> Opcoes -> ImagensM -> Picture

Desenha os menus.

desenhaColecionavel :: [(Colecionavel, Posicao)] -> ImagensC -> [Picture]

Desenha os colecionáveis.

desenhaMapa :: Float -> Float -> [[Bloco]] -> ImagensB -> Pictures

Desenha o mapa.

desenhaLinha :: Float -> Float -> [Bloco] -> ImagensB -> Pictures

Desenha uma linha da matriz que define o mapa.

desenhaBloco :: Float -> Float -> Bloco -> ImagensB -> Pictures

Desenha cada bloco.

desenhaEntidadeAux1 :: [Personagem] -> Personagem -> ImagensP -> CarregaTecla -> ParidadeT -> Pictures

Auxiliar nº1 da desenhaEntidade.

desenhaEntidadeAux2 :: [Personagem] -> ImagensP -> CarregaTecla -> ParidadeT -> Pictures

Auxiliar nº2 da desenhaEntidade.

desenhaEntidade :: Personagem -> ImagensP -> CarregaTecla -> ParidadeT -> Picture

Desenha as Entidades.

reageEvento :: Event -> PrimateKong -> PrimateKong

Função que faz alterações no estado consoante alguns eventos vão acontecendo (clicar em teclas por exemplo).

reageTempo :: Float -> PrimateKong -> PrimateKong

Função que faz alterações no estado consoante o tempo vai passando.

main :: IO ()

Função Main (O que faz o jogo funcionar).

nivel1 :: [[Bloco]]

Mapa do Jogo.