Tarefa1

Description

Módulo para a realização da Tarefa 1 de LI1 em 2023/24.

Documentation

colisoesParede

Arguments

:: Mapa
Mapa do Jogo.

-> Personagem
Personagem em causa.

-> Bool
Valor lógico da colisão entre os limites do mapa / plataformas e o personagem.

A função colisoesParede testa se uma personagem se encontra em colisão com algum dos limites do mapa (topo ou laterais), ou com algum bloco de plataforma. Para nos auxiliar nesta tarefa utilizamos algumas funções que realizam tarefas mais específicas, como por exemplo :

colisaoMapa: Testa se uma personagem colide com algum dos limites do mapa

colisaoBlocos: Testa se uma personagem colide com um bloco de plataforma

hitboxPersonagem: Calcula a hitbox de uma personagem, definifida por um par de posições, sendo estas o canto inferior esquerdo e superior direito

hitboxBlocos: Calcula a hitbox de uma lista de plataformas

posPlataforma: Recebe a matriz de definição do mapa e devolve uma lista com as coordenadas do canto superior esquerdo de cada bloco

Propriedades:

Função principal "colisoesParede":

colisoesParede :: Mapa -> Personagem -> Bool

colisoesParede (Mapa _ _ m) p

| colisoesParede m p || colisaoBlocos (hitboxBlocos $ posPlataforma m) (hitboxPersonagem p) = True

| otherwise = False

Função auxiliar "colisaoMapa":

colisaoMapa :: [[Bloco]] -> Personagem -> Bool

colisaoMapa m (Personagem _ _ (x, y) _ (g, c) _ _ _ _ _)

| x - (g / 2) <= 0 || x + (g / 2) >= th || y - (c / 2) <= 0 || y + (c / 2) >= tv = True

| otherwise = False

where

th = fromIntegral $ length $ head m

tv = fromIntegral $ length m

Função auxiliar "colisaoBlocos":

colisaoBlocos:: [Hitbox] -> Hitbox -> Bool

colisaoBlocos [] _ = False

colisaoBlocos (((xi1, yi1), (xi2, yi2)):xs) ((xe1, yd1), (xe2, yd2)) =

elem True [x >= min xe1 xe2 && x <= max xe1 xe2 && y >= min yd1 yd2 && y <= max yd1 yd2 | (x, y) <- v] || colisaoBlocos xs ((xe1, yd1), (xe2, yd2))

where

v = [(x, y) | x <- [xi1, xi2], y <- [yi1, yi2]]

Função auxiliar "posPlataforma":

posPlataforma :: [[Bloco]] -> [Posicao]

posPlataforma [] = []

posPlataforma m =

concatMap (\(l, bs) -> [(fromIntegral c, fromIntegral l) | (b, c) <- zip bs [0..], b == Plataforma]) $ zip [0..] m

Função auxiliar "hitboxBlocos":

hitboxBlocos :: [Posicao] -> [Hitbox]

hitboxBlocos = map (\(x, y) -> ((x, y + 1), (x + 1, y)))

Função auxiliar "hitboxPersonagem":

hitboxPersonagem :: Personagem -> Hitbox

hitboxPersonagem (Personagem _ _ (x,y) _ (g,c) _ _ _ _ _)

= ((xe1,yd1),(xe2,yd2))

where

xe1 = x - (g / 2)

yd1 = y + (c / 2)

xe2 = x + (g / 2)

yd2 = y - (c / 2)

Exemplos de utilização:

>>> colisoesParedes (Mapa ((0, 0), Este) (4, 4) [[Plataforma, Plataforma, Plataforma], [Plataforma, Vazio, Plataforma], [Plataforma, Plataforma, Plataforma]]) (Personagem (2, 0) Jogador (1, 1) Este (1, 1) False False 5 0 (False, 0))
True

>>> colisoesParedes (Mapa ((0, 0), Este) (4, 3) [[Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio]]) (Personagem (2, 0) Jogador (1, 1) Este (1, 1) False False 5 0 (False, 0))
False

colisaoMapa

Arguments

:: [[Bloco]]
Matriz que define o mapa.

-> Personagem
Personagem em causa.

-> Bool
Valor lógico da colisão do personagem com os limites do mapa.

Função que verifica se um personagem colide com os limites do mapa.

hitboxBlocos

Arguments

:: [Posicao]
Lista de posições de plataformas.

-> [Hitbox]
Lista de hitboxes dos blocos de plataforma.

Função que calcula a hitbox de uma lista de plataformas.

colisaoBlocos

Arguments

:: [Hitbox]
Lista das hitboxes de cada bloco do tipo plataforma.

-> Hitbox
Hitbox do Personagem.

-> Bool
Valor lógico da colisão dos blocos do tipo plataforma com o personagem.

Função que testa se um personagem colide com um bloco do tipo plataforma.

posPlataforma

Arguments

:: [[Bloco]]
Matriz que define o Mapa.

-> [Posicao]
Lista das cooordenadas do canto superior esquerdo do bloco.

Função que recebe a matriz de definição do mapa e devolve uma lista com as coordenadas do canto superior esquerdo de cada bloco de plataforma.

hitboxPersonagem

Arguments

:: Personagem
Personagem em causa.

-> Hitbox
Hitbox gerada.

Função que calcula a hitbox de uma personagem, definifida por um par de posições, sendo estas o canto inferior esquerdo e superior direito, respetivamente

colisoesPersonagens

Arguments

:: Personagem
Personagem 1.

-> Personagem
Personagem 2.

-> Bool
Valor lógico da colisão entre a Personagem 1 e a Personagem 2.

A função colisoesPersonagens testa se dois personagens se encontram em colisão, ou seja, testa se a hitbox de um dos personagens interceta de alguma forma a hitbox do outro Para definir esta função, utilizamos a função hitboxPersonagem a também uma outra função auxiliar :

colisaoHitbox: Verifica se duas hitboxes se intercetam

Propriedades:

Função Principal "colisoesPersonagens":

colisoesPersonagens :: Personagem -> Personagem -> Bool

colisoesPersonagens p1 p2 = colPersona (hitPersona p1) (hitPersona p2)

Função auxiliar "colisaoHitbox"

colisaoHitbox :: Hitbox -> Hitbox -> Bool

colisaoHitbox ((xi1, yi1), (xi2, yi2)) ((xe1, yd1), (xe2, yd2)) =

elem True [x >= min xe1 xe2 && x <= max xe1 xe2 && y >= min yd1 yd2 && y <= max yd1 yd2 | (x, y) <- v]

where

v = [(x, y) | x <- [xi1, xi2], y <- [yi1, yi2]]

Exemplos de utiilização:

>>> colisoesPersonagens (Personagem (2,1) Jogador (1,1) Este (2,2) False False 3 0 (False, 0)) (Personagem (2,1) Jogador (0.5, 0.5) Este (2,2) False False 3 0 (False, 0))
True

>>> colisoesPersonagens (Personagem (2,1) Jogador (1,1) Este (2,2) False False 3 0 (False, 0)) (Personagem (2,1) Jogador (4, 4) Este (2,2) False False 3 0 (False, 0))
False

colisaoHitbox

Arguments

:: Hitbox
Hitbox de um personagem.

-> Hitbox
Hitbox de um outro personagem.

-> Bool
Valor lógico da colisão entre as duas hitboxes.

Função que verifica se duas hitboxes se intersetam.