Tarefa2

Description

Módulo para a realização da Tarefa 2 de LI1 em 2023/24.

Documentation

valida

Arguments

:: Jogo
Definição do Jogo.

-> Bool
Valor lógico da função. Se for True, o Jogo é válido.

O objetivo desta tarefa é implementar a função valida, cujo objetivo é verificar se um dado jogo não viola nenhuma das seguintes restrições:

O mapa tem “chão”, i.e. uma plataforma que impede que o jogador ou outro personagem caia fora do mapa.

Todos os inimigos tẽm a propriedade ressalta a True, enquanto que o jogador a tem a False.

A posição inicial de um jogador não pode colidir com a posição inicial de um outro personagem. Note que as posições iniciais de inimigos podem colidir entre estes.

Número mínimo de inimigos: 2 (dois).

Inimigos Fantasma têm exatamente 1 (uma) vida.

Escadas não podem começar/terminar em alçapões, e pelo menos uma das suas extremidades tem que ser do tipo Plataforma.

Alçapões não podem ser menos largos que o jogador.

Não podem existir personagens nem colecionãveis “dentro” de plataformas ou alçapões, isto é. o bloco (na matriz do mapa) correspendente à posição de um personagem ou objeto tem que ser Vazio.

Na definição da função valida, utilizamos funções auxiliares que verificam o valor lógico (True ou False) de cada uma das restrições acima enumeradas, como se verifica em:

Função temChao, que verifica se um determinado mapa "tem chão", ou seja, uma plataforma que impede que o jogador caia fora do mapa.

Função testeRessalta, que verifica se todos os inimigos têm a propriedade ressalta a True e o jogador a False.

Função colisaoInicial, que verifica se a posição inicial do jogador não colide com a posição inicial de um outro personagem.

Função minInimigos, que verifica se a condição do número mínimo de inimigos no mapa ser 2 (dois) se verifica.

Função verificaVidas, que verifica se todos os inimigos do tipo Fantasma tem exatamente 1 (uma) vida.

Função verificaCB, que verifica se todas as escadas do mapa não começam nem terminam em alçapões e se pelo menos uma das suas extremidades é do tipo Plataforma.

Função verificaAlcapao, que verifica se um jogador é menos largo que um alçapão.

Função posColecPerson, que verifica se algum personagem ou colecionável está "dentro" de um bloco do tipo Vazio do mapa.

No caso da função testeRessalta, utilizamos:

A função retiraRessalta, que retira o valor lógico da propriedade Ressalta de um personagem.

No caso da função colisaoInicial, utilizamos:

A função posicaoPersonagem, que retira a posição de um personagem.

No caso da função verificaVidas, utilizamos:

A função auxiliar verificaVidas, que verifica se todos os inimigos do tipo fantasma têm exatamente 1 (uma) vida. A função principal (verificaVidas) apenas verifica se todos os casos são do tipo True.

No caso da função verificaCB, utilizamos funções auxiliares como:

A função cimaBEscada, que recebe a matriz que define o mapa e a lista das posições de blocos de escada no mapa e devolve uma lista de pares que correspondem ao bloco acima e a baixo da escada.

A função posEscada, que calcula a lista de posições de blocos de escada.

No caso da função posColecPerson, utilizamos as seguintes funções auxiliares:

A função verPosColec, verifica se uma posição dada é válida para um objeto colecionável no mapa.

A função blocoVazio, verifica se um bloco na posição especificada está vazio.

A função dentroPlataformaOuAlcapao, verifica se um bloco na posição especificada é uma plataforma ou alçapão.

Propriedades:

Função principal "valida":

valida :: Jogo -> Bool

valida (Jogo m i c p)

= (all (==True) [f1,f2,f3,f4,f5,f6,f7,f8])

where

f1 = temchao m

f2 = testeRessalta i j

f3 = not (colisaoInicial m i)

f4 = minInimigos i

f5 = verificaVidas (verificaVidasAux i)

f6 = verificaCB (cimaBEscada m (posEscada m))

f7 = verificaAlcapao j

f8 = posColecPerson m i c

Função auxiliar "temChao":

temChao :: Mapa -> Bool

temChao (Mapa _ _ map)

| all (\x -> x == Plataforma) lastRow = True

| otherwise = False

where lastRow = last map

Função auxiliar "testeRessalta":

testeRessalta :: [Personagem] -> Personagem -> Bool

testeRessalta l (Personagem {tipo = j, ressalta = f} = (j == Jogador) && all (==True) (map retiraRessalta l)

where

retiraRessalta :: Personagem -> Bool

retiraRessalta (Personagem _ _ p2 _ _ _ _ _ _ _) = p2

Função auxiliar "colisaoInicial":

colisaoInicial :: Mapa -> [Personagem] -> Bool

colisaoInicial _ [] = False

colisaoInicial (Mapa (pi,_ ) _ _) (p:ps) = if pi == posicaoPersonagem p then True else colisaoInicial  (Mapa (pi, undefined) undefined undefined) ps

where

posicaoPersonagem :: Personagem -> Posicao

posicaoPersonagem (Personagem _ _ p _ _ _ _ _ _ _) = p

Função auxiliar "minInimigos":

minInimigos :: [Personagem] -> Bool

minInimigos li = length li >= 2

Função auxiliar "verificaVidas":

verificaVidas :: [Bool] -> [Bool]

verificaVidas = all (==True)

Função "verificaVidasAux":

verificaVidasAux [] = []

verificaVidasAux ((Personagem _ t _ _ _ _ _ v _ _):ps)

| t == Jogador || t == MacacoMalvado = True: verificaVidasAux ps

| t == Fantasma && v == 1 = True : verificaVidasAux ps

| otherwise = False :verificaVidasAux ps

Função auxiliar "verificaCB":

verificaCB :: [(Bloco, Bloco)] -> Bool

verificaCB [] = True

| x == Plataforma && (y == Plataforma || y == Vazio || y == Escada) = verificaCB r

| (x == Plataforma || x == Vazio || x == Escada) && y == Plataforma = verificaCB r

| otherwise = False

Função auxiliar "cimaBEscada":

cimaBEscada :: Mapa -> [Posicao] -> [(Bloco, Bloco)]

cimaBEscada _ [] = []

cimaBEscada (Mapa (pi, d) pf m) ((x, y) : r) =

let e = cimaBEscada (Mapa (pi, d) pf m) r

cima = if x > 0 then m !! floor (x - 1) !! floor y else Vazio

baixo = if x < (fromIntegral (length m)) -1 then m !! floor (x + 1) !! floor y else Vazio

extremidades = (cima, baixo)

in if cima /= Alcapao && baixo /= Alcapao && (cima == Plataforma || baixo == Plataforma)

then extremidades : e

else e

Função auxiliar "posEscada":

posEscada :: Mapa -> [Posicao]

posEscada (Mapa _ _ m) = concatMap (\(l, bs) -> [(fromIntegral c, fromIntegral l) | (b, c) <- zip bs [0..], b == Escada]) $ zip [0..] m

Função auxiliar "verificaAlcapao":

verificaAlcapao :: Personagem -> Bool

verificaAlcapao (Personagem _ _ _ _ (l,a) _ _ _ _ _) = l <= 1

Função auxiliar "posColecPerson":

posColecPerson :: Mapa -> [Personagem] -> [(Colecionavel, Posicao)] -> Bool

posColecPerson (Mapa _ _ m) personagens colecionaveis =

all (\(colec, pos) -> verPosColec m colec pos) colecionaveis

where

verPosColec :: [[Bloco]] -> Colecionavel -> Posicao -> Bool

verPosColec m Moeda (x, y) =

blocoVazio m (floor x) (floor y) && not (dentroPlatformaOuAlcapao m (floor x) (floor y))

verPosColec m Martelo (x, y) =

blocoVazio m (floor x) (floor y) && not (dentroPlatformaOuAlcapao m (floor x) (floor y))

blocoVazio :: [[Bloco]] -> Int -> Int -> Bool

blocoVazio m x y = m !! x !! y == Vazio

dentroPlatformaOuAlcapao :: [[Bloco]] -> Int -> Int -> Bool

dentroPlatformaOuAlcapao m x y =

case m !! x !! y of

Plataforma -> True

Alcapao -> True

_ -> False

temChao

Arguments

:: Mapa
Mapa do Jogo.

-> Bool
Valor lógico do chão do Mapa ser todo feito por blocos do tipo Plataforma.

Verifica se o mapa tem chão válido. Um mapa é considerado válido se a última linha consistir apenas em blocos do tipo Plataforma.

Exemplos de Utilização:

>>> temChao (Mapa (1,1) (2,2) [[Vazio, Vazio, Plataforma], [Plataforma, Plataforma, Plataforma], [Vazio, Vazio, Vazio]])
True
>>> temChao (Mapa (1,1) (2,2) [[Vazio, Vazio, Plataforma], [Plataforma, Plataforma, Escada], [Vazio, Vazio, Vazio]])
False

testeRessalta

Arguments

:: [Personagem]
Lista de inimigos.

-> Personagem
Jogador.

-> Bool
Se todos os inimigos têm a propriedade ressalta a True, enquanto que o jogador a tem a False, fornecerá o valor lógico True. Se não forem cumpridas estas condições dará False.

Verifica se os inimingos têm a propriedade ressalta a True, enquanto que o jogador tem a False.

colisaoInicial

Arguments

:: Mapa
Mapa do Jogo.

-> [Personagem]
Lista de inimigos.

-> Bool
Valor lógico do local de nascimento dos inimigos não ser igual ao local de "spawn" do Jogador.

Verifica se o jogador começa com colisão.

minInimigos

Arguments

:: [Personagem]
Lista de inimigos.

-> Bool
Valor lógico da lista de inimigos ser maior ou igual a 2.

Verifica se o nº mínimo de inimigos é igual ou superior ao definido (2).

verificaVidas

Arguments

:: [Bool]
Lista de Valores lógicos obtidos da função verificaVidasAux.

-> Bool
Valor lógico que verifica se a condição imposta se observa.

Função que verifica se todos os casos estudados na função verificaVidasAux são verdade.

verificaVidasAux

Arguments

:: [Personagem]
Lista de inimigos.

-> [Bool]
Lista dos valores lógicos de se verificar (um a um) a condição de a vida dos inimigos do tipo Fantasma ser igual 1.

Função que verifica se todos os inimigos do tipo Fantasma têm exatamente uma vida.

posEscada

Arguments

:: Mapa
Mapa do Jogo.

-> [Posicao]
Lista de posições de blocos de escada.

Função que calcula a lista de posições de blocos de escada.

cimaBEscada

Arguments

:: Mapa
Mapa do Jogo.

-> [Posicao]
Lista de posições de blocos de escada.

-> [(Bloco, Bloco)]
Lista de pares que correspondem ao bloco acima e a baixo da escada.

Função que recebe a matriz que define o mapa e a lista das posições de blocos de escada no mapa e devolve uma lista de pares que correspondem ao bloco acima e a baixo da escada.

verificaCB

Arguments

:: [(Bloco, Bloco)]
lista de pares que correspondem ao bloco acima e em baixo das escada

-> Bool
Valor Lógico de se verificarem as duas condições

Função que verifica se as escadas não começam nem terminam em alçapões e se pelo menos uma das suas extremidades tem uma Plataforma.

verificaAlcapao

Arguments

:: Personagem
Jogador

-> Bool
Valor lógico do Personagem do Jogador ser menor que um bloco de Alçapão do mapa

Função que verifica se um jogador é menos largo que um bloco de Alçapão do mapa.

posColecPerson

Arguments

:: Mapa
Mapa do Jogo.

-> [Personagem]
Lista de personagens.

-> [(Colecionavel, Posicao)]
Lista de colecionáveis e suas posições.

-> Bool
True se todos estiverem em blocos vazios e fora de plataformas ou alçapões, False caso contrário.

Função que verifica se algum personagem ou colecionável está "dentro" de um bloco do tipo Vazio do mapa.

verPosColec

Arguments

:: [[Bloco]]
Matriz do mapa.

-> Colecionavel
Tipo de colecionável.

-> Posicao
Posição do colecionável.

-> Bool
True se a posição é válida, False caso contrário.

Verifica se um colecionável está "dentro" de um bloco do tipo Vazio do mapa. Para colecionáveis do tipo Moeda, a posição é considerada válida se o bloco correspondente estiver vazio e não dentro de uma plataforma ou alçapão. Para colecionáveis do tipo Martelo, as mesmas condições se aplicam como para Moeda.

Exemplos de utilização:

>>> verPosColec [[]] Moeda (0, 0)
False

>>> verPosColec [[Vazio, Plataforma], [Vazio, Vazio]] Martelo (1, 0)
False

blocoVazio

Arguments

:: [[Bloco]]
Matriz do mapa.

-> Int
Coordenada x da posição.

-> Int
Coordenada y da posição.

-> Bool
True se o bloco está vazio, False caso contrário.

Verifica se um bloco na posição especificada está vazio.

Exemplo de utilização:

>>> blocoVazio [[Vazio, Plataforma], [Vazio, Vazio]] 0 0
True

dentroPlatformaOuAlcapao

Arguments

:: [[Bloco]]
Matriz do mapa.

-> Int
Coordenada x da posição.

-> Int
Coordenada y da posição.

-> Bool
True se o bloco é uma plataforma ou alçapão, False caso contrário.

Verifica se um bloco na posição especificada é uma plataforma ou alçapão.

Exemplo de utilização:

>>> dentroPlatformaOuAlcapao [[Vazio, Plataforma], [Vazio, Vazio]] 0 1
True

testColecionaveis :: [(Colecionavel, Posicao)]

jogadorTeste :: Personagem

inimigoTeste :: Personagem

personagensTeste :: [Personagem]

personagensTeste1 :: [Personagem]

exemploMapaDiferente :: Mapa

exemploMapaTrue :: Mapa

exemploMapaFalse :: Mapa