Tarefa3

Description

Módulo para a realização da Tarefa 3 de LI1 em 2023/24.

Documentation

movimenta

Arguments

:: Semente
Semente do Jogo.

-> Tempo
Tempo do Jogo.

-> Jogo
Jogo.

-> Jogo
Jogo após atualização.

Função que irá animar todas as personagens, i.e. calcular as suas novas posições e respectivas consequências.

A função movimenta deve animar todas as personagens, i.e. calcular as suas novas posições e respectivas consequências. Para isso, a função tem de testar uma lista não recursiva de acontecimentos e condições a que os personagens estão sujeitos. Para concretizar o que dissemos anteriormente, aplicamos algumas funções auxiliares que têm como objetivo verificar determinadas condições e realizar certas mudanças no jogo, como é o caso das seguintes funções:

inPerdeVida: Função que retira as vidas aos inimigos que colidem com a hitbox de dano de um jogador (martelo).

filtraInimigos: Função que filtra os inimigos com 1 vida numa lista de inimigos.

pisaAlcapao: Função que atualiza a matriz do mapa do jogo quando um alçapão é pisado.

gravPersonagem: Função que faz os personagens cair (mudar a velocidade) quando não estão numa plataforma.

jogLevaDano: Função que tira uma vida ao jogador se ele entrar em colisão com um inimigo.

apanhaColec: Função que atualiza o jogo quando o jogador obtém um colecionável (aumentando a pontuação ou permitindo dar dano).

naoAtravessa: Função que impede o personagem de sair do mapa e de atravessar blocos de plataforma (diminui a velocidade do personagem para 0).

Propriedades:

Função principal "movimenta":

movimenta :: Semente -> Tempo -> Jogo -> Jogo

movimenta s t j@(Jogo m@(Mapa c1 c2 c3) i c p) = apanhaColec $ Jogo mn (filtraInimigos (inPerdeVida i p)) c (jogLevaDano i (naoAtravessa m (gravPersonagem mn p)))

where mn = (Mapa c1 c2 (pisaAlcapao m p))

Função auxiliar "inPerdeVida":

inPerdeVida:: [Personagem] -> Personagem -> [Personagem]

inPerdeVida [] _ = []

inPerdeVida ((Personagem c1 c2 c3 c4 c5 c6 c7 v c9 (a,t)):ls) p

| hitboxPersonagem (Personagem c1 c2 c3 c4 c5 c6 c7 v c9 (a,t)) == hitboxDanoJ p && (pDaDano p == True) = ((Personagem c1 c2 c3 c4 c5 c6 c7 (v - 1) c9 (a,t))) : inPerdeVida ls p

| otherwise = (Personagem c1 c2 c3 c4 c5 c6 c7 v c9 (a,t)) : inPerdeVida ls p

Função "pDaDano":

pDaDano :: Personagem -> Bool

pDaDano (Personagem _ _ _ _ _ _ _ v _ (a, t))

| v > 0 && a == True = True

| otherwise          = False

Função "hitboxDanoJ":

hitboxDanoJ :: Personagem -> Hitbox

hitboxDanoJ p@(Personagem _ _ _ d (l, a) _ _ v _ (ad, t))

| pDaDano p && d == Este  = ((x2,y1), (x2+1,y2))

| pDaDano p && d == Oeste = ((x1-1,y1), (x1,y2))

| otherwise = ((0,0),(0,0))

where ((x1,y1), (x2,y2)) = hitboxPersonagem p

Função auxiliar "filtraInimigos":

filtraInimigos :: [Personagem] -> [Personagem]

filtraInimigos l = filter (\(Personagem _ _ _ _ _ _ _ v _ _) -> v == 1) l

Função auxiliar "pisaAlcapao":

pisaAlcapao :: Mapa -> Personagem -> [[Bloco]]

pisaAlcapao (Mapa _ _ m) p@(Personagem _ t (x, y) _ _ _ _ _ _ _)

| t == Jogador && any (==(fst (hitboxPersonagem p))) (posAlcapao m) = substituiBMatriz m (xp,yp) Vazio

| otherwise = m

where (xp,yp) = fst (hitboxPersonagem p)

Função "substituiBMatriz":

substituiBMatriz :: [[Bloco]] -> Posicao -> Bloco -> [[Bloco]]

substituiBMatriz [] _ _ = []

substituiBMatriz (h:t) (0,y) b = (substituiBloco h y b) : t

substituiBMatriz (h:t) (x,y) b = h : substituiBMatriz t (x-1,y) b

Função "substituiBloco":

substituiBloco :: [Bloco] -> Double -> Bloco -> [Bloco]

substituiBloco [] _ _= []

substituiBloco (h:t) 0 b = b:t

substituiBloco (h:t) n b = h : substituiBloco t (n-1) b

Função "posAlcapao":

posAlcapao :: [[Bloco]] -> [Posicao]

posAlcapao [] = []

posAlcapao m =  concatMap (\(l, bs) -> [(fromIntegral c, fromIntegral l) | (b, c) <- zip bs [0..], b == Alcapao]) $ zip [0..] m

Função auxiliar "gravPersonagem":

gravPersonagem :: Mapa -> Personagem -> Personagem

gravPersonagem (Mapa _ _ m) (Personagem (vx, vy) c2 (x, y) c4 c5 c6 c7 c8 c9 c10)

| any (==(fst (hitboxPersonagem p))) (posPlataforma m) = p

| otherwise = (Personagem (vx, vy + 10) c2 (x, y) c4 c5 c6 c7 c8 c9 c10)

where p = (Personagem (vx, vy) c2 (x, y) c4 c5 c6 c7 c8 c9 c10)

Função auxiliar "jogLevaDano":

jogLevaDano :: [Personagem] -> Personagem -> Personagem

jogLevaDano [] p = p

jogLevaDano l p@(Personagem _ _ _ _ _ _ _ v _ _)

| any (==True) (colisaoInJog l p) = p {vida = (v-1)}

| otherwise = p

Função "colisaoInJog":

colisaoInJog :: [Personagem] -> Personagem -> [Bool]

colisaoInJog [] p = [False]

colisaoInJog (h:t) p = colisoesPersonagens h p : colisaoInJog t p

Função auxiliar "apanhaColec":

apanhaColec :: Jogo -> Jogo

apanhaColec (Jogo m i ((colec,pos):r) p@(Personagem _ _ _ _ _ _ _ _ pontuacao _))

| colisaoHitbox (hitboxPersonagem p) (hitboxColecionavel pos) && colec == Moeda   = (Jogo m i r p {pontos = pontuacao + 100})

| colisaoHitbox (hitboxPersonagem p) (hitboxColecionavel pos) && colec == Martelo = (Jogo m i r p {aplicaDano = (True,10)})

| otherwise = (Jogo m i ((colec,pos):r) p)

Função "hitboxColecionavel":

hitboxColecionavel :: Posicao -> Hitbox

hitboxColecionavel (x,y) = ((x1,y1),(x2,y2)))

where

x1 = x - (1/2)

y1 = y + (1/2)

x2 = x + (1/2)

y2 = y - (1/2)

Função principal "naoAtravessa":

naoAtravessa :: Mapa -> Personagem -> Personagem

naoAtravessa m@(Mapa _ _ mb) p@(Personagem (vx,vy) _ (x,y) _ (l,a) _ _ _ _ _)

| c1        = p {velocidade = (0, vy)}

| c2        = p {velocidade = (vx, 0)}

| c1 && c2  = p {velocidade = (0,0)}

| otherwise = p

where c1 = colisoesParede m p

c2 = any (==(fst (hitboxPersonagem p))) (posPlataforma mb)

Exemplos de utilização:

>>> movimenta (fromIntegral 1) 2 (Jogo (Mapa ((2.5, 1.5), Este) (5, 5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Alcapao, Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]]) ([(Personagem (1, 0) Fantasma (2.5, 1.5) Oeste (1, 1) False True 1 0 (False, 0))]) ([(Moeda, (1.5,1.5)), (Martelo,(4.5,4.5))]) (Personagem (0, 0) Jogador (1.5, 1.5) Este (1, 1) False True 3 0 (True, 5)))
Jogo Mapa ((2.5, 1.5), Este) (5, 5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]] [] [(Martelo,(4.5,4.5))] Personagem (0, 10) Jogador (1.5, 1.5) Este (1, 1) False True 2 1 (True, 5)

>>> movimenta (fromIntegral 1) 2 (Jogo (Mapa ((2.5, 1.5), Este) (5, 5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Plataforma, Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]]) ([(Personagem (1, 0) Fantasma (3.5, 3.5) Oeste (1, 1) False True 1 0 (False, 0))]) ([(Martelo,(4.5,4.5))]) (Personagem (0, 0) Jogador (1.5, 1.5) Este (1, 1) False True 3 0 (False, 0)))
Jogo Mapa ((2.5, 1.5), Este) (5, 5) [[Plataforma, Plataforma, Plataforma, Plataforma, Plataforma], [Vazio, Vazio, Vazio, Vazio, Vazio], [Vazio, Plataforma, Vazio, Vazio, Vazio], [Vazio, Vazio, Vazio, Vazio, Vazio], [Plataforma, Plataforma, Plataforma, Plataforma, Plataforma]] [(Personagem (1, 0) Fantasma (3.5, 3.5) Oeste (1, 1) False True 1 0 (False, 0))] [(Martelo,(4.5,4.5))] Personagem (0, 0) Jogador (1.5, 1.5) Este (1, 1) False True 3 0 (False, 0)

movimentoJ

Arguments

:: Personagem
Personagem do Jogador

-> Personagem
Personagem do Jogador atualizado

movimentoI

Arguments

:: [Personagem]
Lista de inimigos do Jogo

-> [Personagem]
Lista atualizadas dos inimigos do Jogo

inPerdeVida

Arguments

:: [Personagem]
Lista de Inimigos do mapa.

-> Personagem
Jogador.

-> [Personagem]
Lista de inimigos do mapa que perderam vidas.

Função que retira as vidas aos inimigos que colidem com a hitbox de dano de um jogador (martelo).

pDaDano

Arguments

:: Personagem
Personagem do Jogador.

-> Bool
Valor lógico da propriedade "aplicaDano" estar ativa.

Função que verifica se um personagem pode dar dano.

hitboxDanoJ

Arguments

:: Personagem
Jogador.

-> Hitbox
Hitbox de Dano (Martelo).

Função que determina a hitbox de dano do Jogador (Martelo).

filtraInimigos

Arguments

:: [Personagem]
Lista de inimigos

-> [Personagem]
Lista de inimigos com uma vida

Função que filtra os inimigos com 1 vida numa determinada lista de inimigos.

pisaAlcapao

Arguments

:: Mapa
Mapa do Jogo.

-> Personagem
Jogador.

-> [[Bloco]]
Matriz do Mapa com os alçapões atualizados (após serem pisados).

Função que atualiza a matriz do mapa quando um jogador pisa um alçapão.

substituiBMatriz

Arguments

:: [[Bloco]]
Matriz do Mapa.

-> Posicao
Posição de um bloco no Mapa.

-> Bloco
Bloco.

-> [[Bloco]]
Matriz com o bloco atualizado.

Função que substitui um bloco por outro numa matriz (mapa).

substituiBloco

Arguments

:: [Bloco]
Lista de Blocos.

-> Double
Posição na lista de blocos.

-> Bloco
Bloco.

-> [Bloco]
Lista de Blocos atualizada.

Função que substitui um bloco por outro numa lista de blocos (linha na matriz do mapa).

posAlcapao

Arguments

:: [[Bloco]]
Matriz do Mapa.

-> [Posicao]
Posição dos blocos do tipo Alcapao.

Função que calcula a posição de todos os blocos de alçapão no Mapa.

gravPersonagem

Arguments

:: Mapa
Mapa do Jogo.

-> Personagem
Jogador.

-> Personagem
Jogador com a velocidade atualizada.

Função que faz os personagens cair (mudar a velocidade) quando não estão numa plataforma.

jogLevaDano

Arguments

:: [Personagem]
Lista de inimigos.

-> Personagem
Jogador.

-> Personagem
Jogador com a vida atualizada.

Função que tira uma vida ao jogador se ele entrar em colisão com um inimigo.

colisaoInJog

Arguments

:: [Personagem]
Lista de inimigos.

-> Personagem
Jogador.

-> [Bool]
Lista dos valores lógicos dos inimigos estarem ou não em contato com o Jogador.

Função que verifica se algum inimigo se encontra em colisão com o jogador.

apanhaColec

Arguments

:: Jogo
Jpgo.

-> Jogo
Jogo após os colecionáveis serem colectados (atualizado).

Função que atualiza o jogo quando o jogador obtém um colecionável (aumentando a pontuação ou permitindo dar dano).

hitboxColecionavel

Arguments

:: Posicao
Posição de um colecionável no Mapa.

-> Hitbox
Hitbox do colecionável dado.

Função que calcula a hitbox de um colécionável.

naoAtravessa

Arguments

:: Mapa
Mapa do Jogo.

-> Personagem
Jogador.

-> Personagem
Jogador com a velocidade atualizada.

Função que impede o personagem de sair do mapa e de atravessar blocos de plataforma (diminui a velocidade do personagem para 0).

aleatorioTodos

Arguments

:: Mapa
Mapa do Jogo

-> Semente
Semente Aleatória

-> [Personagem]
Lista de Inimigos

-> [Personagem]
Lista de Inimigos atualizada

função eu gera movimento aleatório a uma lista de inimigos

aleatorioUm

Arguments

:: Mapa
Mapa do Jogo

-> Personagem
Inimigo qualquer

-> Int
Número de 19 dígitos

-> Personagem
Inimigo atualizado

naoAndaDireita2

Arguments

:: [[Bloco]]
Matriz que define o Mapa do Jogo.

-> Personagem
Personagem.

-> Bool
Valor lógico de o personagem não se poder deslocar para a direita.

naoAndaEsquerda2

Arguments

:: [[Bloco]]
Matriz que define o Mapa do Jogo.

-> Personagem
Personagem.

-> Bool
Valor lógico de o personagem não se poder deslocar para a esquerda.

Função que verifica se um personagem não se pode deslocar para a esquerda.

posVazio2

Arguments

:: [[Bloco]]
Matriz que define o Mapa do Jogo.

-> [Posicao]
Lista de posições de todos os blocos vazios na matriz que define o Mapa do Jogo.

numaEscada2

Arguments

:: Mapa
Mapa do Jogo.

-> Personagem
Personagem.

-> Bool
Valor lógico de o personagem estar em colisão com um bloco de Escada (pode subir ou descer).