TarefaComplementar

Description

Módulo para a realização da Tarefa 6 de LI1 em 2023/24.

Documentation

data CarregaTecla

Constructors

Car

Fields

acarregar :: Bool

acao :: Maybe Acao

teclaa :: Bool

teclad :: Bool

teclaw :: Bool

type ParidadeT = Bool

movimentoTecla :: Jogo -> CarregaTecla -> Jogo

Movimento tecla.

podeSaltar2 :: Mapa -> Personagem -> Bool

Verifica se o personagem pode saltar.

atualizaPersonagem2 :: Mapa -> Acao -> Personagem -> Personagem

Função que atualiza as personagens.

atualizaJogador2 :: Mapa -> Maybe Acao -> Personagem -> Personagem

Função que atualiza o Jogador.

atualiza2 :: [Maybe Acao] -> Maybe Acao -> Jogo -> Jogo

perderJogo :: Jogo -> Bool

Função que verifica se o jogador perde o Jogo.

posPer :: Personagem -> Posicao

Posição de um personagem.

arredonda :: Posicao -> (Int, Int)

Arredonda para o Integer mais próximo.

mudaTipo :: (Int, Int) -> Posicao

Muda de Pares de Int para posição.

matrizPMapa :: Posicao -> Posicao

Função que passa de coordenadas dentro da matriz para o mapa real

cantoCentro :: Posicao -> Posicao

Função que passa das coordenadas do canto superior esquerdo do mapa para as coordenadas do centro da matriz

posPlataformas :: [[Bloco]] -> [Posicao]

Posição das plataformas.

posVazios :: [[Bloco]] -> [Posicao]

Posição dos vazios.

posEscadas :: [[Bloco]] -> [Posicao]

Posicao das escadas.

posAlcapoes :: [[Bloco]] -> [Posicao]

Posição dos alcapoes.

colisaoLado :: Mapa -> Personagem -> Personagem

Colisão lateral

colisaoLadoI :: Mapa -> [Personagem] -> [Personagem]

Colisão lateral inimigos

acabaMartelo :: Personagem -> Personagem

Função que altera o tempo de duração do Martelo até este acabar.

maisColecionavel :: (Colecionavel, Posicao) -> Jogo -> Jogo

Adiciona colecionáveis

posJogColec :: Posicao -> Posicao -> Bool

Posição do Jogador e do colecionável intersetam-se.

apanhaColec2 :: Jogo -> Jogo

Função que verifica se os colecionáveis são coletados e aplica os seus respetivos efeitos.

apanhaEstrela :: Jogo -> Bool

Função que verifica se o jogador apanhou a estrela.

gravidadeJ2 :: Mapa -> Personagem -> Personagem

Função que trata de aplicar a gravidade ao jogador.

gravidadeI2 :: Mapa -> Personagem -> Personagem

Função que trata de aplicar a gravidade aos inimigos fantasmas.

gravidadeI2T :: Mapa -> [Personagem] -> [Personagem]

Gravidade aplicada à lista de inimigos.

numaEscadaI :: [[Bloco]] -> Personagem -> Bool

Função que verifica se um personagem se encontra numa escada.

movimentoJ2 :: Personagem -> Personagem

Movimento do Jogador

movimentoI2 :: [Personagem] -> [Personagem]

Movimento dos inimigos

arredondaPos :: Posicao -> [Posicao] -> Posicao

Arredonda os valores das posições.

distancia :: Posicao -> Posicao -> Double

Verifica a distância entre dois pontos.

movAleatorioIni :: Mapa -> Semente -> [Personagem] -> [Personagem]

O movimento dos inimigos é aleatório.

proMovimento :: Mapa -> Personagem -> Personagem

Função que calcula o próximo movimento do inimigo

colisaoLadosI :: Mapa -> Personagem -> (Bool, Bool)

Colisão com os lados dos inimigos

inPerdeVida2 :: [Personagem] -> Personagem -> [Personagem]

Inimigos perdem vida.

hitboxDano2 :: Personagem -> [Posicao]

Hitboxes do dano.

jogadorLevaD2 :: [Personagem] -> Personagem -> Personagem

Jogador leva dano.

colisoesDano2 :: [Personagem] -> Personagem -> [Bool]

Colisões com dano.

colisoesPersonagens2 :: Personagem -> Personagem -> Bool

Colisões entre personagens.

movimenta3 :: Semente -> Tempo -> Jogo -> Jogo

Função movimenta versão 3

pisaAlcapao2 :: Mapa -> Personagem -> [[Bloco]]

Verifica se um alçapão foi pisado.

numaEscada3 :: [[Bloco]] -> Personagem -> Personagem

Verifica se está numa escada versão 3

corEscada :: [[Bloco]] -> Personagem -> Personagem

Verifa se um inimigo está numa escada.

corEscadaAux :: [[Bloco]] -> [Posicao]

Auxiliar da função acima.

corEscadaI :: [[Bloco]] -> [Personagem] -> [Personagem]

Aplica a função corEscada à lista de inimigos.

vazioEmBaixo :: [[Bloco]] -> Personagem -> Bool

Verifica se numa posição dada é vazio.

nascInicial :: Mapa -> [Personagem] -> Bool

Verifica onde as personagens dão spawn.

passaAlcapao :: Personagem -> Bool

verifica se o jogador passa pelo alçapão.

colisaoCoisas2 :: [Personagem] -> Personagem -> [(Colecionavel, Posicao)] -> Mapa -> Bool

Verifica se algum colecionavel colide com algum bloco.

posInicialTudo :: [Personagem] -> Personagem -> [(Colecionavel, Posicao)] -> [Posicao]

Posição inicial dos personagens e dos colecionáveis.

cimaBEscada2 :: Mapa -> [Posicao] -> [(Bloco, Bloco)]

Faz o mesmo que a função cimaBEscada da Tarefa2.

valida2 :: Jogo -> Bool

Função valida versão 2.

verificaParidade :: Tempo -> ParidadeT

Verifica se o tempo é par ou ímpar (isto entra na movimentação do macaco).